# contracts/v1_api_swagger.json
{
  "openapi": "3.0.1",
...
  "info": {
    "title": "Ouroboros Agent API",
    "description": "The Ouroboros Agent API is used to allow agents to connect to the Ouroboros server.",
    "version": "1.3",
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://www.mozilla.org/en-US/MPL/2.0/"
    }
  },
  "servers": [
    {
      "url": "https://{defaultHost}",
      "description": "The production server",
      "variables": {
        "defaultHost": {
          "default": "www.example.com"
        }
      }
    },
    {
      "url": "http://{hostAddress}:{hostPort}",
      "description": "The insecure server",
      "variables": {
        "hostAddress": {
          "default": "localhost"
        },
        "hostPort": {
          "default": "8080"
        }
      }
    }
  ],
  "security": [
    {
      "bearer_auth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorObject": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The id of the task"
          },
          "attack_id": {
            "type": "integer",
            "format": "int64",
            "description": "The id of the attack"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "description": "The time the task was started"
          },
          "status": {
            "type": "string",
            "description": "The status of the task"
          },
          "skip": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "The offset of the keyspace"
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "The limit of the keyspace"
          }
        },
        "required": [
          "id",
          "attack_id",
          "start_date",
          "status"
        ]
      },
...
    }
  },
  "paths": {
    "/api/v1/client/agents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "get": {
        "summary": "Gets an instance of an agent",
        "tags": [
          "Agents"
        ],
        "description": "Returns an agent",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "getAgent",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 2571,
                      "host_name": "weissnat",
                      "client_signature": "macOS, High Sierra (10.13)",
                      "operating_system": "linux",
                      "state": "active",
                      "devices": [
                        "cpu",
                        "GPU"
                      ],
                      "advanced_configuration": {
                        "agent_update_interval": 36,
                        "use_native_hashcat": false,
                        "backend_device": null,
                        "opencl_devices": null,
                        "enable_additional_hash_types": false
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Updates the agent",
        "tags": [
          "Agents"
        ],
        "description": "Updates an agent",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "updateAgent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 2573,
                      "host_name": "stoltenberg",
                      "client_signature": "Linux, RHEL 6.10",
                      "operating_system": "linux",
                      "state": "active",
                      "devices": [
                        "cpu",
                        "GPU"
                      ],
                      "advanced_configuration": {
                        "agent_update_interval": 24,
                        "use_native_hashcat": false,
                        "backend_device": null,
                        "opencl_devices": null,
                        "enable_additional_hash_types": false
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "The id of the agent"
                  },
                  "host_name": {
                    "type": "string",
                    "description": "The hostname of the agent"
                  },
                  "client_signature": {
                    "type": "string",
                    "description": "The signature of the client"
                  },
                  "operating_system": {
                    "type": "string",
                    "description": "The operating system of the agent"
                  },
                  "devices": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "The descriptive name of a GPU or CPU device."
                    }
                  }
                },
                "required": [
                  "id",
                  "host_name",
                  "client_signature",
                  "operating_system",
                  "devices"
                ]
              },
              "examples": {
                "request_example_1": {
                  "summary": "A request example",
                  "value": {
                    "id": 2573,
                    "client_signature": "Linux, RHEL 6.10",
                    "enabled": true,
                    "last_ipaddress": "",
                    "last_seen_at": null,
                    "host_name": "stoltenberg",
                    "operating_system": "linux",
                    "token": "jwTK41L3PsXctHdcJL7Pt1zt",
                    "user_id": 12821,
                    "created_at": "2024-10-01T21:16:40.622-04:00",
                    "updated_at": "2024-10-01T21:16:40.622-04:00",
                    "devices": [
                      "cpu",
                      "GPU"
                    ],
                    "advanced_configuration": {
                      "agent_update_interval": 24,
                      "use_native_hashcat": false,
                      "backend_device": null,
                      "opencl_devices": null,
                      "enable_additional_hash_types": false
                    },
                    "state": "active",
                    "custom_label": null
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/client/tasks/new": {
      "get": {
        "summary": "Request a new task from server",
        "tags": [
          "Tasks"
        ],
        "description": "Request a new task from the server, if available.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "getNewTask",
        "responses": {
          "204": {
            "description": "no new task available"
          },
          "200": {
            "description": "new task available",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 802,
                      "attack_id": 1003,
                      "start_date": "2024-10-01T21:16:42.896-04:00",
                      "status": "pending",
                      "skip": 0,
                      "limit": 0
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "get": {
        "summary": "Request the task information",
        "tags": [
          "Tasks"
        ],
        "description": "Request the task information from the server.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "getTask",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "id": 803,
                      "attack_id": 1004,
                      "start_date": "2024-10-01T09:13:42.205-04:00",
                      "status": "pending",
                      "skip": 0,
                      "limit": 0
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Record not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/tasks/{id}/submit_crack": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "post": {
        "summary": "Submit a cracked hash result for a task",
        "tags": [
          "Tasks"
        ],
        "description": "Submit a cracked hash result for a task.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "sendCrack",
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "message": "Hash cracked successfully, 1 hashes remaining, task running."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No more uncracked hashes"
          },
          "404": {
            "description": "Hash value not found",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Hash not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HashcatResult"
              },
              "examples": {
                "request_example_1": {
                  "summary": "A request example",
                  "value": {
                    "timestamp": "2024-10-01T21:16:43.171-04:00",
                    "hash": "dummy_hash_2",
                    "plain_text": "dummy_plain"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/client/tasks/{id}/submit_status": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "post": {
        "summary": "Submit a status update for a task",
        "tags": [
          "Tasks"
        ],
        "description": "Submit a status update for a task. This includes the status of the current guess and the devices.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "sendStatus",
        "parameters": [],
        "responses": {
          "204": {
            "description": "status received successfully"
          },
          "202": {
            "description": "status received successfully, but stale"
          },
          "410": {
            "description": "status received successfully, but task paused"
          },
          "422": {
            "description": "malformed status data",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Guess not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Record not found"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Bad credentials"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStatus"
              }
            }
          },
          "required": true,
          "description": "status"
        }
      }
    },
    "/api/v1/client/tasks/{id}/accept_task": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer",
            "format": "int64"
          },
          "required": true,
          "description": "id"
        }
      ],
      "post": {
        "summary": "Accept Task",
        "tags": [
          "Tasks"
        ],
        "description": "Accept an offered task from the server.",
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "operationId": "setTaskAccepted",
        "responses": {
          "204": {
            "description": "task accepted successfully"
          },
          "422": {
            "description": "task already completed",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Task already completed"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          },
          "404": {
            "description": "task not found for agent",
            "content": {
              "application/json": {
                "examples": {
                  "test_example": {
                    "value": {
                      "error": "Record not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorObject"
                }
              }
            }
          }
        }
      }
    },
...
  }
}